requirements.txt
Flask
google-generativeai
python-dotenv
langchain
langchain_experimental
langchain_google_genai
pydantic
flask-cors